Finished Pagination, sort and Filter

+++ server/controllers/categoryController.js
@@ -1,4 +1,5 @@
 const Category = require('../models/category')
+const Products = require('../models/product')
 
 const categoryCrtl = {
   getCategories: async (req, res)=>{
@@ -25,6 +26,11 @@ const categoryCrtl = {
   },
   deleteCategory: async( req, res) => {
     try {
+      const products =  await Products.findOne({category: req.params.id})
+      if(products) return res.status(400).json({
+        msg:"Please delete all products wih a relationship"
+      })
+
       await Category.findByIdAndDelete(req.params.id)
       res.json({msg: "Deleted a Category"})
     } catch (error) {
